+++ Radiantium.Offline/Distribution2D.cs
@@ -11,7 +11,7 @@ namespace Radiantium.Offline
         readonly Distribution1D[] _conditional;//条件分布
         readonly Distribution1D _marginal;//边缘分布
 
-        public Distribution2D(float[] data, int nu, int nv)
+        public Distribution2D(double[] data, int nu, int nv)
         {
             if (nu * nv != data.Length)
             {
@@ -20,9 +20,9 @@ namespace Radiantium.Offline
             _conditional = new Distribution1D[nv];
             for (int v = 0; v < nv; v++) //计算条件概率
             {
-                _conditional[v] = new Distribution1D(new Span<float>(data, v * nu, nu));
+                _conditional[v] = new Distribution1D(new Span<double>(data, v * nu, nu));
             }
-            float[] marginal = new float[nv];
+            double[] marginal = new double[nv];
             for (int v = 0; v < nv; v++) //计算边缘分布
             {
                 marginal[v] = _conditional[v].Integration;
@@ -32,19 +32,19 @@ namespace Radiantium.Offline
 
         public Vector2 SampleContinuous(Vector2 u, out float pdf, out (int, int) offset)
         {
-            float d1 = _marginal.SampleContinuous(u.Y, out float pdf1, out int vj);
-            float d0 = _conditional[vj].SampleContinuous(u.X, out float pdf0, out int vi);
-            pdf = pdf0 * pdf1;
+            double d1 = _marginal.SampleContinuous(u.Y, out double pdf1, out int vj);
+            double d0 = _conditional[vj].SampleContinuous(u.X, out double pdf0, out int vi);
+            pdf = (float)(pdf0 * pdf1);
             offset = (vj, vi);
-            return new(d0, d1);
+            return new((float)d0, (float)d1);
         }
 
         public Vector2 SampleContinuous(Vector2 u, out float pdf)
         {
-            float d1 = _marginal.SampleContinuous(u.Y, out float pdf1, out int vi);
-            float d0 = _conditional[vi].SampleContinuous(u.X, out float pdf0, out _);
-            pdf = pdf0 * pdf1;
-            return new(d0, d1);
+            double d1 = _marginal.SampleContinuous(u.Y, out double pdf1, out int vi);
+            double d0 = _conditional[vi].SampleContinuous(u.X, out double pdf0, out _);
+            pdf = (float)(pdf0 * pdf1);
+            return new((float)d0, (float)d1);
         }
 
         public Vector2 SampleContinuous(Vector2 u)
@@ -56,7 +56,7 @@ namespace Radiantium.Offline
         {
             int iu = Math.Clamp((int)(u * _conditional[0].Count), 0, _conditional.Length - 1);
             int iv = Math.Clamp((int)(v * _marginal.Count), 0, _marginal.Count - 1);
-            return _conditional[iv].Pdf[iu] / _marginal.Integration;
+            return (float)(_conditional[iv].Pdf[iu] / _marginal.Integration);
         }
     }
 }
